sys/ztimer: work around bug in old LLVM
The sum of an `uint16_t` variable and a literal `1` should still be of
type `uin16_t`. And at least with LLVM 16 this seems to be the case,
but not with the LLVM version in our build container :/

So cast what should have been `uint16_t` anyway explicitly to `uint16_t`
to make that buggy version of LLVM happy.

+++ sys/ztimer/core.c
@@ -57,7 +57,7 @@ static bool _ztimer_acquire(ztimer_clock_t *clock)
     unsigned state = irq_disable();
 
     DEBUG("ztimer_acquire(): %p: %" PRIu16 " user(s)\n",
-          (void *)clock, clock->users + 1);
+          (void *)clock, (uint16_t)(clock->users + 1));
 
     if (clock->users++ == 0) {
         if (clock->ops->start) {
@@ -93,7 +93,7 @@ bool ztimer_release(ztimer_clock_t *clock)
     assert(clock->users > 0);
 
     DEBUG("ztimer_release(): %p: %" PRIu16 " user(s)\n",
-          (void *)clock, clock->users - 1);
+          (void *)clock, (uint16_t)(clock->users - 1));
 
     if (--clock->users == 0) {
         /* make sure the timer isn't armed before turning off */
